% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{calc_A}
\alias{calc_A}
\title{Calculate photosynthesis}
\usage{
calc_A(
  T_air = 25,
  PPFD = 1000,
  Patm = 101.325,
  E,
  u = 2,
  leaf_width = 0.01,
  RH = 60,
  Ca = 420,
  Jmax = 100,
  Vcmax = 50
)
}
\arguments{
\item{T_air}{Air temperature, deg C}

\item{PPFD}{Photosynthetic photon flux density, mu mol m-2 s-1}

\item{Patm}{Atmospheric pressure, kPa}

\item{E}{Transpiration rate, mmol m-2 s-1}

\item{u}{Wind speed above the leaf boundary layer, m s-1}

\item{leaf_width}{Leaf width, m}

\item{RH}{Relative humidity, in \%}

\item{Ca}{Atmospheric CO2 concentration (ppm)}

\item{Jmax}{Maximum rate of electron transport at 25 deg C (mu mol m-2 s-1)}

\item{Vcmax}{Maximum carboxylation rate at 25 deg C (mu mol m-2 s-1)}
}
\value{
Photosynthetic rate (mol m-2 s-1)
}
\description{
Calculate photosynthesis
}
\examples{
# Calculate A for a scalar value of E
calc_A(E = 10)

# Calculate leaf VPD along transpiration supply stream
Weibull = fit_Weibull() # Fit Weibull parameters
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c) # Calculate Pcrit based on Weibull curve

P = Ps_to_Pcrit(Pcrit = Pcrit) # Create Ps to Pcrit vector
trans = trans_from_vc(P = P, b = b, c = c) # Create vector of transpiration supply stream

calc_A(E = trans)
}
