% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_gain.R
\name{calc_costgain}
\alias{calc_costgain}
\title{Calculate all costs and gains}
\usage{
calc_costgain(
  P,
  b = -2.5,
  c = 2,
  Amax_gross = NULL,
  Amax_net = NULL,
  kmax_25 = 4,
  Tair = 25,
  VPD = 1.5,
  ratiocrit = 0.05,
  PPFD = 1000,
  Patm = 101.325,
  Wind = 2,
  Wleaf = 0.01,
  LeafAbs = 0.5,
  Tcrit = 50,
  T50 = 51,
  Ca = 420,
  Jmax = 100,
  Vcmax = 50,
  constant_kmax = FALSE,
  Rd0 = 0.92,
  TrefR = 25,
  ...
)
}
\arguments{
\item{P}{Vector of equally spaced water potentials ranging from Ps to Pcrit,
-MPa}

\item{b}{Weibull scale parameter}

\item{c}{Weibull shape parameter}

\item{Amax_gross}{Maximum potential gross photosynthetic rate (A) that A values are normalized
by to obtain the carbon gain. If NULL, this is the instantaneous maximum
potential rate described by Sperry et al. (2016).}

\item{Amax_net}{Maximum potential net photosynthetic rate (A) that A values are normalized
by to obtain the carbon gain. If NULL, this is the instantaneous maximum
potential rate described by Sperry et al. (2016).}

\item{kmax_25}{Max plant conductance at 25 deg C, mmol s-1 m-2 MPa-1}

\item{Tair}{Air temperature, deg C}

\item{VPD}{Air vapor pressure deficit, kPa}

\item{ratiocrit}{Percentage of maximum conductivity at which hydraulic damage
is considered irreversible}

\item{PPFD}{Photosynthetic photon flux density, mu mol m-2 s-1}

\item{Patm}{Atmospheric pressure, kPa}

\item{Wind}{Wind speed above the leaf boundary layer, m s-1}

\item{Wleaf}{Leaf width, m}

\item{LeafAbs}{Leaf absorptance of solar radiation (0-1)}

\item{Tcrit}{Leaf temperature at which F0(T) transitions from slow- to fast- rise, deg C}

\item{T50}{Leaf temperature midway between Tcrit and Tmax, deg C}

\item{Ca}{Atmospheric CO2 concentration (ppm)}

\item{Jmax}{Maximum rate of electron transport at 25 deg C (mu mol m-2 s-1)}

\item{Vcmax}{Maximum carboxylation rate at 25 deg C (mu mol m-2 s-1)}

\item{constant_kmax}{TRUE if the kmax does not vary with temperature for
simulations; else FALSE}

\item{Rd0}{Day respiration rate at reference temperature (TrefR). Must be a positive value.}

\item{TrefR}{Reference temperature for Rd (deg C)}

\item{...}{Further parameters passed to \code{plantecophys::Photosyn}.}
}
\value{
A data frame with columns "P", "ID" and "cost_gain", where "P" denotes
    the leaf water potential, "ID" is one of CG, HC, or TC, and "cost_gain" is the
    corresponding value.
}
\description{
Calculate the carbon gain, hydraulic cost, and thermal costs.
}
\examples{
# Calculate leaf VPD along transpiration supply stream
Weibull = fit_Weibull() # Fit Weibull parameters
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c) # Calculate Pcrit based on Weibull curve
P = Ps_to_Pcrit(Pcrit = Pcrit) # Create Ps to Pcrit vector

calc_costgain(P, b, c)
}
