% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{calc_Dleaf}
\alias{calc_Dleaf}
\title{Vapor pressure deficit}
\usage{
calc_Dleaf(T_leaf = 25, T_air = 25, RH = 60)
}
\arguments{
\item{T_leaf}{Leaf temperature, deg C}

\item{T_air}{Air temperature, deg C}

\item{RH}{Relative humidity, in \%}
}
\value{
Leaf vapor pressure deficit, kPa
}
\description{
Vapor pressure deficit
}
\examples{
# Calculate Tleaf for a given scalar value of leaf temperature
calc_Dleaf(T_leaf = 30)


# Calculate leaf VPD along transpiration supply stream
Weibull = fit_Weibull() # Fit Weibull parameters
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c) # Calculate Pcrit based on Weibull curve

P = Ps_to_Pcrit(Pcrit = Pcrit) # Create Ps to Pcrit vector
trans = trans_from_vc(P = P, b = b, c = c) # Create vector of transpiration supply stream
Tleaves = calc_Tleaf(E = trans) # Get leaf temperatures

Dleaves = calc_Dleaf(T_leaf = Tleaves)
}
