% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{calc_gHa}
\alias{calc_gHa}
\title{Boundary layer conductance to forced convection}
\usage{
calc_gHa(T_air = 25, u = 2, Patm = 101.325, leaf_width = 0.01)
}
\arguments{
\item{T_air}{Air temperature, deg C}

\item{u}{Wind speed above the leaf boundary layer, m s-1}

\item{Patm}{Atmospheric pressure, kPa}

\item{leaf_width}{Leaf width, m}
}
\value{
Boundary layer conductance to forced convection, mol m-2 s-1
}
\description{
Boundary layer conductance to forced convection
}
