% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{calc_gr}
\alias{calc_gr}
\title{Radiative conductance}
\usage{
calc_gr(T_air = 25, epsilon = 0.97)
}
\arguments{
\item{T_air}{Air temperature, deg C}

\item{epsilon}{Leaf emissivity}
}
\value{
Radiative conductance, mol m-2 s-1
}
\description{
Radiative conductance
}
\examples{
calc_gr(T_air = 30)
}
