% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{calc_gw}
\alias{calc_gw}
\title{Stomatal conductance}
\usage{
calc_gw(
  E,
  Tleaf = 25,
  Patm = 101.325,
  Tair = 25,
  VPD = 1.5,
  PPFD = 1000,
  Wind = 8,
  Wleaf = 0.01
)
}
\arguments{
\item{E}{Transpiration rate, mmol m-2 s-1}

\item{Tleaf}{Leaf temperature, deg C}

\item{Patm}{Atmospheric pressure, kPa}

\item{Tair}{Air temperature, deg C}

\item{VPD}{Air vapor pressure deficit, kPa}

\item{PPFD}{Photosynthetic photon flux density, mu mol m-2 s-1}

\item{Wind}{Wind speed above the leaf boundary layer, m s-1}

\item{Wleaf}{Leaf width, m}
}
\value{
Stomatal conductance to water vapor, mol m-2 s-1
}
\description{
Stomatal conductance
}
\examples{
# Calculate leaf conductance for a scalar value of E
calc_gw(E = 3)

# Calculate gw along transpiration supply stream
Weibull = fit_Weibull() # Fit Weibull parameters
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c) # Calculate Pcrit based on Weibull curve

P = Ps_to_Pcrit(Pcrit = Pcrit) # Create Ps to Pcrit vector
trans = trans_from_vc(P = P, b = b, c = c) # Create vector of transpiration supply stream
Tleaves = calc_Tleaf(E = trans) # Get leaf temperatures
Dleaves = plantecophys::VPDairToLeaf(Tleaf = Tleaves, Tair = 25, VPD = 1.5) # Get leaf VPD

calc_gw(E = trans, Tleaf = Tleaves)
}
