% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulics.R
\name{calc_kmax}
\alias{calc_kmax}
\title{Temperature dependence of conductivity}
\usage{
calc_kmax(kmax_25 = 4, Tair = 25, constant_kmax = FALSE)
}
\arguments{
\item{kmax_25}{Max plant conductance at 25 deg C, mmol s-1 m-2 MPa-1}

\item{Tair}{Air temperature, deg C}

\item{constant_kmax}{TRUE if the kmax does not vary with temperature for
simulations; else FALSE}
}
\value{
Whole plant conductivity at Tair, mmol s-1 m-2 MPa-1
}
\description{
Calculate whole plant conductivity as a function of air
    temperature
}
