% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulics.R
\name{calc_Pcrit}
\alias{calc_Pcrit}
\title{Solve for Pcrit}
\usage{
calc_Pcrit(b = -2.5, c = 2, ratiocrit = 0.05)
}
\arguments{
\item{b}{Weibull scale parameter}

\item{c}{Weibull shape parameter}

\item{ratiocrit}{Percentage of maximum conductivity at which hydraulic damage
is considered irreversible}
}
\value{
The critical leaf water potential (-MPa)
}
\description{
Calculate the critical leaf water potential
}
\examples{
Weibull = fit_Weibull()
b = Weibull[1,1]
c = Weibull[1,2]

Pcrit = calc_Pcrit(b, c)

# Change the critical ration
Pcrit1 = calc_Pcrit(b, c, 0.025)
}
