% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atm.R
\name{calc_VPD}
\alias{calc_VPD}
\title{Vapor pressure deficit}
\usage{
calc_VPD(T_air = 25, RH = 60)
}
\arguments{
\item{T_air}{Air temperature, deg C}

\item{RH}{Relative humidity, in \%}
}
\value{
Vapor pressure deficit, kPa
}
\description{
Calculate the atmospheric VPD at given relative humidity and air
    temperature
}
