% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal_damage.R
\name{F0_func}
\alias{F0_func}
\title{F0-T Curve}
\usage{
F0_func(Tleaf, Tcrit = 50, T50 = 51, F0_max = 1000, F0_min = 500)
}
\arguments{
\item{Tleaf}{Leaf temperature, deg C}

\item{Tcrit}{Leaf temperature at which F0(T) transitions from slow- to fast- rise, deg C}

\item{T50}{Leaf temperature midway between Tcrit and Tmax, deg C}

\item{F0_max}{Maximum F0 (achieved at Tmax)}

\item{F0_min}{Minimum (i.e. undamaged) F0}
}
\value{
Minimum fluorescence (F0)
}
\description{
Calculates the value of minimum fluorescence (F0) associated
    with a given leaf temperature. F0 values serve as a proxy of thermal
    damage to Photsystem II.
}
\examples{
Tleaf = seq(25, 50, length.out = 500)
curve = F0_func(Tleaf = Tleaf)
plot(Tleaf, curve, type = "l")
}
