% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulics.R
\name{fit_Weibull}
\alias{fit_Weibull}
\title{Fit Weibull}
\usage{
fit_Weibull(P50 = 4.9, P88 = 6.41)
}
\arguments{
\item{P50}{LWP at 50 PLC (-MPa)}

\item{P88}{LWP at 88 PLC (-MPa)}
}
\value{
A data frame containing the Weibull parameters b and c
}
\description{
Fit Weibull parameters given values of P50 and P88
}
