% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_gain.R
\name{gain_min_costs}
\alias{gain_min_costs}
\title{Calculate gain minus costs}
\usage{
gain_min_costs(
  P,
  b = -2.5,
  c = 2,
  Amax_gross = NULL,
  Amax_net = NULL,
  kmax_25 = 4,
  T_air = 25,
  ratiocrit = 0.05,
  PPFD = 1000,
  RH = 90,
  Patm = 101.325,
  u = 2,
  leaf_width = 0.01,
  Tcrit = 50,
  T50 = 51,
  Ca = 420,
  Jmax = 100,
  Vcmax = 50,
  constant_kmax = FALSE,
  Rd0 = 0.92,
  TrefR = 25
)
}
\arguments{
\item{P}{Vector of equally spaced water potentials ranging from Ps to Pcrit,
-MPa}

\item{b}{Weibull scale parameter}

\item{c}{Weibull shape parameter}

\item{Amax_gross}{Maximum potential gross photosynthetic rate (A) that A values are normalized
by to obtain the carbon gain. If NULL, this is the instantaneous maximum
potential rate described by Sperry et al. (2016).}

\item{Amax_net}{Maximum potential net photosynthetic rate (A) that A values are normalized
by to obtain the carbon gain. If NULL, this is the instantaneous maximum
potential rate described by Sperry et al. (2016).}

\item{kmax_25}{Max plant conductance at 25 deg C, mmol s-1 m-2 MPa-1}

\item{T_air}{Air temperature, deg C}

\item{ratiocrit}{Percentage of maximum conductivity at which hydraulic damage
is considered irreversible}

\item{PPFD}{Photosynthetic photon flux density, mu mol m-2 s-1}

\item{RH}{Relative humidity, in \%}

\item{Patm}{Atmospheric pressure, kPa}

\item{u}{Wind speed above the leaf boundary layer, m s-1}

\item{leaf_width}{Leaf width, m}

\item{Tcrit}{Leaf temperature at which F0(T) transitions from slow- to fast- rise, deg C}

\item{T50}{Leaf temperature midway between Tcrit and Tmax, deg C}

\item{Ca}{Atmospheric CO2 concentration (ppm)}

\item{Jmax}{Maximum rate of electron transport at 25 deg C (mu mol m-2 s-1)}

\item{Vcmax}{Maximum carboxylation rate at 25 deg C (mu mol m-2 s-1)}

\item{constant_kmax}{TRUE if the kmax does not vary with temperature for
simulations; else FALSE}

\item{Rd0}{Day respiration rate at reference temperature (TrefR). Must be a positive value.}

\item{TrefR}{Reference temperature for Rd (deg C)}
}
\value{
A data frame with columns "P", "ID" and "CG_min_C", where "P" denotes
    the leaf water potential, "ID" is either HC or CC, and "CG_min_C" is the
    corresponding value of carbon gain minus the cost identified in "ID".
}
\description{
Calculates the difference between the carbon gain and either
    hydraulic cost or summed hydraulic and thermal costs.
}
\examples{
# Calculate leaf VPD along transpiration supply stream
Weibull = fit_Weibull() # Fit Weibull parameters
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c) # Calculate Pcrit based on Weibull curve
P = Ps_to_Pcrit(Pcrit = Pcrit) # Create Ps to Pcrit vector

gain_min_costs(P, b, c)
}
