% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_gain.R
\name{hydraulic_cost}
\alias{hydraulic_cost}
\title{Hydraulic cost function}
\usage{
hydraulic_cost(
  P,
  b = -2.5,
  c = 2,
  kmax_25 = 4,
  T_air = 25,
  ratiocrit = 0.05,
  constant_kmax = FALSE
)
}
\arguments{
\item{P}{Vector of equally spaced water potentials ranging from Ps to Pcrit,
-MPa}

\item{b}{Weibull scale parameter}

\item{c}{Weibull shape parameter}

\item{kmax_25}{Max plant conductance at 25 deg C, mmol s-1 m-2 MPa-1}

\item{T_air}{Air temperature, deg C}

\item{ratiocrit}{Percentage of maximum conductivity at which hydraulic damage
is considered irreversible}

\item{constant_kmax}{TRUE if the kmax does not vary with temperature for
simulations; else FALSE}
}
\value{
Hydraulic cost, unitless
}
\description{
Calculates the normalized hydraulic cost as described in Sperry
    et al. 2016
}
\examples{
# Calculate leaf VPD along transpiration supply stream
Weibull = fit_Weibull() # Fit Weibull parameters
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c) # Calculate Pcrit based on Weibull curve
P = Ps_to_Pcrit(Pcrit = Pcrit) # Create Ps to Pcrit vector

# Calculate hydraulic cost
hydraulic_cost(P, b, c)
}
