% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulics.R
\name{Ps_to_Pcrit}
\alias{Ps_to_Pcrit}
\title{Leaf water potential vector}
\usage{
Ps_to_Pcrit(Ps = 0, Pcrit = 4, pts = 500)
}
\arguments{
\item{Ps}{Soil water potential, -MPa}

\item{Pcrit}{Critical leaf water potential, -MPa}

\item{pts}{Number of steps}
}
\value{
A vector of equally-spaced leaf water potentials ranging from the
   soil water potential to the critical water potential
}
\description{
Create vector of equally-spaced steps ranging from soil water
    potential to the critical water potential
}
\examples{
# Calculate Pcrit based on Weibull curve
Weibull = fit_Weibull()
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c)

# Create Ps to Pcrit vector
P = Ps_to_Pcrit(Pcrit = Pcrit)
}
