% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_gain.R
\name{respiratory_cost}
\alias{respiratory_cost}
\title{Respiratory cost function}
\usage{
respiratory_cost(
  P,
  b = -2.5,
  c = 2,
  Amax = NULL,
  kmax_25 = 4,
  T_air = 25,
  PPFD = 1000,
  Patm = 101.325,
  u = 2,
  leaf_width = 0.01,
  RH = 60,
  constant_kmax = FALSE,
  Rd0 = 0.92,
  TrefR = 25
)
}
\arguments{
\item{P}{Vector of equally spaced water potentials ranging from Ps to Pcrit,
-MPa}

\item{b}{Weibull scale parameter}

\item{c}{Weibull shape parameter}

\item{Amax}{Maximum potential photosynthetic rate (A) that A values are normalized
by to obtain the carbon gain. If NULL, this is the instantaneous maximum
potential rate described by Sperry et al. (2016).}

\item{kmax_25}{Max plant conductance at 25 deg C, mmol s-1 m-2 MPa-1}

\item{T_air}{Air temperature, deg C}

\item{PPFD}{Photosynthetic photon flux density, mu mol m-2 s-1}

\item{Patm}{Atmospheric pressure, kPa}

\item{u}{Wind speed above the leaf boundary layer, m s-1}

\item{leaf_width}{Leaf width, m}

\item{RH}{Relative humidity, in \%}

\item{constant_kmax}{TRUE if the kmax does not vary with temperature for
simulations; else FALSE}

\item{Rd0}{Day respiration rate (mu mol m-2 s-1) at reference temperature
(TrefR). Must be a positive value.}

\item{TrefR}{Reference temperature for Rd (deg C)}
}
\value{
Respiratory cost, unitless
}
\description{
Calculates the normalized respiratory cost along the
    transpiration supply stream.
}
\examples{
# Calculate leaf water potential along transpiration supply stream
Weibull = fit_Weibull() # Fit Weibull parameters
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c) # Calculate Pcrit based on Weibull curve
P = Ps_to_Pcrit(Pcrit = Pcrit) # Create Ps to Pcrit vector

# Calculate respiratory cost
respiratory_cost(P, b, c)
}
