% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_gain.R
\name{thermal_cost}
\alias{thermal_cost}
\title{Thermal cost function}
\usage{
thermal_cost(
  P,
  b = -2.5,
  c = 2,
  kmax_25 = 4,
  T_air = 25,
  PPFD = 1000,
  RH = 90,
  Patm = 101.325,
  u = 2,
  leaf_width = 0.01,
  Tcrit = 50,
  T50 = 51,
  constant_kmax = FALSE
)
}
\arguments{
\item{P}{Vector of equally spaced water potentials ranging from Ps to Pcrit,
-MPa}

\item{b}{Weibull scale parameter}

\item{c}{Weibull shape parameter}

\item{kmax_25}{Max plant conductance at 25 deg C, mmol s-1 m-2 MPa-1}

\item{T_air}{Air temperature, deg C}

\item{PPFD}{Photosynthetic photon flux density, mu mol m-2 s-1}

\item{RH}{Relative humidity, in \%}

\item{Patm}{Atmospheric pressure, kPa}

\item{u}{Wind speed above the leaf boundary layer, m s-1}

\item{leaf_width}{Leaf width, m}

\item{Tcrit}{Leaf temperature at which F0(T) transitions from slow- to fast- rise, deg C}

\item{T50}{Leaf temperature midway between Tcrit and Tmax, deg C}

\item{constant_kmax}{TRUE if the kmax does not vary with temperature for
simulations; else FALSE}
}
\value{
Normalized thermal cost, unitless
}
\description{
Calculates the normalized thermal cost based on F0-T curve
}
\examples{
# Calculate leaf VPD along transpiration supply stream
Weibull = fit_Weibull() # Fit Weibull parameters
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c) # Calculate Pcrit based on Weibull curve
P = Ps_to_Pcrit(Pcrit = Pcrit) # Create Ps to Pcrit vector

thermal_cost(P, b, c)
}
