% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal_damage.R
\name{TJmax_updated}
\alias{TJmax_updated}
\title{Modified Jmax-T response}
\usage{
TJmax_updated(Tleaf, EaJ, delsJ, EdVJ, Tcrit = 43.4, T50 = 49.6)
}
\arguments{
\item{Tleaf}{Leaf temperature (deg C)}

\item{EaJ, EdVJ, delsJ}{Jmax temperature response parameters}

\item{Tcrit}{Leaf temperature at which F0(T) transitions from slow- to fast- rise, deg C}

\item{T50}{Leaf temperature midway between Tcrit and Tmax, deg C}
}
\value{
Jmax (normalised to Jmax at 25 deg C)
}
\description{
Modified temperature response of Jmax. Based on the peaked Arrhenius temperature
response, but multiplied by (1 - TC) so that Jmax decreases as the TC increases.
}
