% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VarEnv_example}
\alias{VarEnv_example}
\title{Example input data}
\format{
A data frame with 51 rows and 15 variables:
}
\usage{
VarEnv_example
}
\description{
An example input file for running multiple simulations with variable
environmental factors. All inputs are theoretical.
}
\keyword{datasets}
