% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atm.R
\name{vpsat}
\alias{vpsat}
\title{Saturation vapor pressure at a given temperature}
\usage{
vpsat(T = 25)
}
\arguments{
\item{T}{Temperature, deg C}
}
\value{
Saturation vapor pressure, kPa
}
\description{
Saturation vapor pressure at a given temperature
}
