% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulics.R
\name{vulnerability_curve}
\alias{vulnerability_curve}
\title{Vulnerability curve}
\usage{
vulnerability_curve(P, b = -2.5, c = 2)
}
\arguments{
\item{P}{Vector with water potentials, -MPa}

\item{b}{Weibull scale parameter}

\item{c}{Weibull shape parameter}
}
\value{
Vulnerability curve
}
\description{
Calculate the vulnerability curve for a given soil and critical
    water potential
}
\examples{
# Calculate Pcrit based on Weibull curve
Weibull = fit_Weibull()
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c)

# Create Ps to Pcrit vector
P = Ps_to_Pcrit(Pcrit = Pcrit)

# Fit vulnerability curve
vulnerability_curve(P, b, c)
}
