% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_gain.R
\name{C_gain}
\alias{C_gain}
\title{Carbon gain}
\usage{
C_gain(
  P,
  b = -2.5,
  c = 2,
  kmax_25 = 4,
  T_air = 25,
  PPFD = 1000,
  Patm = 101.325,
  u = 2,
  leaf_width = 0.01,
  RH = 60,
  Ca = 420,
  Jmax = 100,
  Vcmax = 50
)
}
\arguments{
\item{P}{Vector of equally spaced water potentials ranging from Ps to Pcrit,
-MPa}

\item{b}{Weibull scale parameter}

\item{c}{Weibull shape parameter}

\item{kmax_25}{Max plant conductance at 25 deg C, mmol s-1 m-2 MPa-1}

\item{T_air}{Air temperature, deg C}

\item{PPFD}{Photosynthetic photon flux density, mu mol m-2 s-1}

\item{Patm}{Atmospheric pressure, kPa}

\item{u}{Wind speed above the leaf boundary layer, m s-1}

\item{leaf_width}{Leaf width, m}

\item{RH}{Relative humidity, in \%}

\item{Ca}{Atmospheric CO2 concentration (ppm)}

\item{Jmax}{Maximum rate of electron transport at 25 deg C (mu mol m-2 s-1)}

\item{Vcmax}{Maximum carboxylation rate at 25 deg C (mu mol m-2 s-1)}
}
\value{
Normalized carbon gain
}
\description{
Calculates the normalized carbon gain as described in Sperry et
    al. 2016
}
\examples{
# Calculate leaf VPD along transpiration supply stream
Weibull = fit_Weibull() # Fit Weibull parameters
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c) # Calculate Pcrit based on Weibull curve
P = Ps_to_Pcrit(Pcrit = Pcrit) # Create Ps to Pcrit vector

C_gain(P, b, c)
}
