% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{calc_A}
\alias{calc_A}
\title{Calculate photosynthesis}
\usage{
calc_A(
  Tair = 25,
  VPD = 1.5,
  PPFD = 1000,
  Patm = 101.325,
  E = 2,
  Wind = 2,
  Wleaf = 0.01,
  LeafAbs = 0.5,
  Ca = 420,
  Jmax = 100,
  Vcmax = 50,
  net = FALSE,
  Rd0 = 0.92,
  TrefR = 25,
  netOrig = TRUE,
  ...
)
}
\arguments{
\item{Tair}{Air temperature, deg C}

\item{VPD}{Air vapor pressure deficit, kPa}

\item{PPFD}{Photosynthetic photon flux density, mu mol m-2 s-1}

\item{Patm}{Atmospheric pressure, kPa}

\item{E}{Transpiration rate, mmol m-2 s-1}

\item{Wind}{Wind speed above the leaf boundary layer, m s-1}

\item{Wleaf}{Leaf width, m}

\item{LeafAbs}{Leaf absorptance of solar radiation (0-1)}

\item{Ca}{Atmospheric CO2 concentration (ppm)}

\item{Jmax}{Maximum rate of electron transport at 25 deg C (mu mol m-2 s-1)}

\item{Vcmax}{Maximum carboxylation rate at 25 deg C (mu mol m-2 s-1)}

\item{net}{TRUE if desired output is net photosynthesis; FALSE if desired output
is gross photosynthesis}

\item{Rd0}{Day respiration rate at reference temperature (TrefR). Must be a positive value.}

\item{TrefR}{Reference temperature for Rd (deg C)}

\item{netOrig}{TRUE if net photosynthesis is to be calculated within
\code{plantecophys::Photosyn}. FALSE if Anet is to be calculated as Agross - Rd.}

\item{\dots}{Further parameters passed to \code{plantecophys::Photosyn}.}
}
\value{
Photosynthetic rate (mol m-2 s-1)
}
\description{
Calculate photosynthesis
}
\examples{
# Calculate A for a scalar value of E
calc_A(E = 10)

# Calculate leaf VPD along transpiration supply stream
Weibull = fit_Weibull() # Fit Weibull parameters
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c) # Calculate Pcrit based on Weibull curve

P = Ps_to_Pcrit(Pcrit = Pcrit) # Create Ps to Pcrit vector
trans = trans_from_vc(P = P, b = b, c = c) # Create vector of transpiration supply stream

calc_A(E = trans)
}
