% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{calc_gHa}
\alias{calc_gHa}
\title{Boundary layer conductance to forced convection}
\usage{
calc_gHa(Tair = 25, Wind = 2, Patm = 101.325, Wleaf = 0.025)
}
\arguments{
\item{Tair}{Air temperature, deg C}

\item{Wind}{Wind speed above the leaf boundary layer, m s-1}

\item{Patm}{Atmospheric pressure, kPa}

\item{Wleaf}{Leaf width, m}
}
\value{
Boundary layer conductance to forced convection, mol m-2 s-1
}
\description{
Boundary layer conductance to forced convection
}
