% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{calc_Rd}
\alias{calc_Rd}
\title{Daytime respiration rate}
\usage{
calc_Rd(
  Tair = 25,
  VPD = 1.5,
  PPFD = 1000,
  E = 2,
  Wind = 2,
  Patm = 101.325,
  Wleaf = 0.01,
  LeafAbs = 0.5,
  Rd0 = 0.92,
  TrefR = 25
)
}
\arguments{
\item{Tair}{Air temperature, deg C}

\item{VPD}{Air vapor pressure deficit, kPa}

\item{PPFD}{Photosynthetic photon flux density, mu mol m-2 s-1}

\item{E}{Transpiration rate, mmol m-2 s-1}

\item{Wind}{Wind speed above the leaf boundary layer, m s-1}

\item{Patm}{Atmospheric pressure, kPa}

\item{Wleaf}{Leaf width, m}

\item{LeafAbs}{Leaf absorptance of solar radiation (0-1)}

\item{Rd0}{Day respiration rate (mu mol m-2 s-1) at reference temperature
(TrefR). Must be a positive value.}

\item{TrefR}{Reference temperature for Rd (deg C)}
}
\value{
Rd (mu mol m-2 s-1)
}
\description{
Calculates the daytime respiration rate (Rd) with the Heskel et
    al. 2016 global polynomial model.
}
\examples{
# Calculate daytime respiration for scalar value of E
calc_Rd(E = 10)


# Calculate leaf temperature along transpiration supply stream
Weibull = fit_Weibull() # Fit Weibull parameters
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c) # Calculate Pcrit based on Weibull curve

P = Ps_to_Pcrit(Pcrit = Pcrit) # Create Ps to Pcrit vector
E = trans_from_vc(P = P, b = b, c = c) # Create vector of transpiration supply stream

Rd_vect = calc_Rd(E = E) # Get leaf temperatures
}
