% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atm.R
\name{calc_Rnet}
\alias{calc_Rnet}
\title{Calculate net radiation}
\usage{
calc_Rnet(Tair = 25, PPFD = 1000, VPD = 1.5, albedo = 0.15, epsilon = 0.97)
}
\arguments{
\item{Tair}{Air temperature, deg C}

\item{PPFD}{Photosynthetic photon flux density, mu mol m-2 s-1}

\item{VPD}{Atmospheric vapor pressure deficit, kPa}

\item{albedo}{Leaf albedo, unitless}

\item{epsilon}{Leaf emissivity, unitless}
}
\value{
Net radiation, W m-2
}
\description{
Calculate net radiation
}
