% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{calc_Tleaf}
\alias{calc_Tleaf}
\title{Leaf temperature}
\usage{
calc_Tleaf(
  T_air = 25,
  PPFD = 1000,
  RH = 60,
  E,
  u = 2,
  Patm = 101.325,
  leaf_width = 0.01
)
}
\arguments{
\item{T_air}{Air temperature, deg C}

\item{PPFD}{Photosynthetic photon flux density, mu mol m-2 s-1}

\item{RH}{Relative humidity, in \%}

\item{E}{Transpiration rate, mmol m-2 s-1}

\item{u}{Wind speed above the leaf boundary layer, m s-1}

\item{Patm}{Atmospheric pressure, kPa}

\item{leaf_width}{Leaf width, m}
}
\value{
Leaf temperature, deg C
}
\description{
Leaf temperature
}
\examples{
# Calculate leaf temperature for scalar value of E
calc_Tleaf(E = 10)


# Calculate leaf temperature along transpiration supply stream
Weibull = fit_Weibull() # Fit Weibull parameters
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c) # Calculate Pcrit based on Weibull curve

P = Ps_to_Pcrit(Pcrit = Pcrit) # Create Ps to Pcrit vector
trans = trans_from_vc(P = P, b = b, c = c) # Create vector of transpiration supply stream

Tleaves = calc_Tleaf(E = trans) # Get leaf temperatures
}
