% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{calc_Tleaf}
\alias{calc_Tleaf}
\title{Leaf temperature}
\usage{
calc_Tleaf(
  Tair = 25,
  VPD = 1.5,
  E = 0.2,
  PPFD = 1000,
  Wind = 2,
  Patm = 101.325,
  Wleaf = 0.025,
  LeafAbs = 0.5
)
}
\arguments{
\item{Tair}{Air temperature, deg C}

\item{VPD}{Air vapor pressure deficit, kPa}

\item{E}{Transpiration rate, mmol m-2 s-1}

\item{PPFD}{Photosynthetic photon flux density, mu mol m-2 s-1}

\item{Wind}{Wind speed above the leaf boundary layer, m s-1}

\item{Patm}{Atmospheric pressure, kPa}

\item{Wleaf}{Leaf width, m}

\item{LeafAbs}{Leaf absorptance of solar radiation (0-1)}
}
\value{
Leaf temperature, deg C
}
\description{
Calculate leaf temperature from energy balance. Adapted from
    '\code{plantecophys::FindTleaf}.
}
\examples{
# Calculate leaf temperature for scalar value of E
calc_Tleaf(E = 0.2, Tair = 25)


# Calculate leaf temperature along transpiration supply stream
Weibull = fit_Weibull() # Fit Weibull parameters
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c) # Calculate Pcrit based on Weibull curve

P = Ps_to_Pcrit(Pcrit = Pcrit) # Create Ps to Pcrit vector
trans = trans_from_vc(P = P, b = b, c = c) # Create vector of transpiration supply stream

Tleaves = calc_Tleaf(E = trans) # Get leaf temperatures

}
