% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulics.R
\name{trans_from_vc}
\alias{trans_from_vc}
\title{Transpiration supply function}
\usage{
trans_from_vc(
  P,
  kmax_25 = 0.5,
  Tair = 25,
  b = -2.5,
  c = 2,
  constant_kmax = FALSE
)
}
\arguments{
\item{P}{Vector with water potentials, -MPa}

\item{kmax_25}{Max plant conductance at 25 deg C, mmol s-1 m-2 MPa-1}

\item{Tair}{Air temperature, deg C}

\item{b}{Weibull scale parameter}

\item{c}{Weibull shape parameter}

\item{constant_kmax}{TRUE if the kmax does not vary with temperature for
simulations; else FALSE}
}
\value{
Transpiration, mmol s-1 m-2
}
\description{
Calculate transpiration as integral of a vulnerability curve
}
\examples{
# Calculate Pcrit based on Weibull curve
Weibull = fit_Weibull()
b = Weibull[1,1]
c = Weibull[1,2]
Pcrit = calc_Pcrit(b, c)

# Create Ps to Pcrit vector
P = Ps_to_Pcrit(Pcrit = Pcrit)

# Create vector of transpiration supply stream
trans = trans_from_vc(P = P, b = b, c = c)
}
